Extract location service tier
